Module Patdiff_kernel

module Ansi_output : sig ... end

module Ascii_output : sig ... end

module Compare_core : sig ... end

module Comparison_result : sig ... end

module Configuration : sig ... end

module Diff_input : sig ... end

module File_helpers : sig ... end

module File_name : sig ... end

Used to determine which name to use for a file, depending on the operation.

module Float_tolerance : sig ... end

module Format : sig ... end

Patdiff_format is the home of all the internal representations of the formatting that will be applied to the diff. ie. prefixes, suffixes, & valid styles.

module Html_output : sig ... end

module Hunks : sig ... end

module Import : sig ... end

module Is_binary : sig ... end

Heuristics for detecting files that are clearly binary.

module Output : sig ... end

module Patdiff_core : sig ... end

module Should_keep_whitespace : sig ... end

Heuristics for whitespace-sensitive inputs (e.g., if it looks like Python).