Module Patdiff_kernel.Float_tolerance

val apply : string Import.Patience_diff.Hunks.t -> Core.Percent.t -> context:int -> string Import.Patience_diff.Hunks.t

apply hunks tolerance ~context converts diff ranges into context if the diff looks like floating point numbers that changed by less than tolerance. If necessary, narrows or splits the resulting hunks to limit context to context lines.

Float tolerance must be applied before refinement.