Module Patdiff_kernel.Format

Patdiff_format is the home of all the internal representations of the formatting that will be applied to the diff. ie. prefixes, suffixes, & valid styles.

module Color : sig ... end

module Style : sig ... end

module Rule : sig ... end

A rule consists of a styled prefix, a styled suffix, and a style. Rules are applied to strings using functions defined in Output_ops.

module Rules : sig ... end

Rules are configured in the configuration file. Default values are provided in Configuration.

module Location_style : sig ... end