Module Format.Color

module RGB6 : sig ... end

module Gray24 : sig ... end

type t =

| Black

| Red

| Green

| Yellow

| Blue

| Magenta

| Cyan

| White

| Default

| Gray

| Bright_black

| Bright_red

| Bright_green

| Bright_yellow

| Bright_blue

| Bright_magenta

| Bright_cyan

| Bright_white

| RGB6 of RGB6.t

| Gray24 of Gray24.t

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t -> t -> int

include Ppx_quickcheck_runtime.Quickcheckable.S with type t := t

val quickcheck_generator : t Base_quickcheck.Generator.t

val quickcheck_observer : t Base_quickcheck.Observer.t

val quickcheck_shrinker : t Base_quickcheck.Shrinker.t

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

include Core.Comparable.S with type t := t

include Base.Comparable.S with type t := t

include Base.Comparisons.S with type t := t

include Base.Comparisons.Infix with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : t -> min:t -> max:t -> t

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : t -> min:t -> max:t -> t Base.Or_error.t

include Base.Comparator.S with type t := t

type comparator_witness

val comparator : ( t, comparator_witness ) Base.Comparator.comparator

val validate_lbound : min:t Core.Maybe_bound.t -> t Validate.check

val validate_ubound : max:t Core.Maybe_bound.t -> t Validate.check

val validate_bound : min:t Core.Maybe_bound.t -> max:t Core.Maybe_bound.t -> t Validate.check

module Replace_polymorphic_compare : Base.Comparable.Polymorphic_compare with type t := t

module Map : Core.Map.S with type Key.t = t with type Key.comparator_witness = comparator_witness

module Set : Core.Set.S with type Elt.t = t with type Elt.comparator_witness = comparator_witness

val rgb6_exn : (int * int * int) -> t

rgb6_exn r g b and gray24_exn level raise if the values are out of bound.

val gray24_exn : int -> t